+++ TheSultansOfSuggestion/Assets/Scripts/Player/PlayerAttack/PlayerAttack.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 abstract public class PlayerAttack : MonoBehaviour
 {
-    protected int damage = 2;
-    public int GetDamage()
+    protected float damage = 2;
+    public float GetDamage()
     {
         return this.damage;
     }
